requirements.txt
fastapi
uvicorn[standard]
pypdf
python-multipart
mistralai
pymupdf
numpy
python-dotenv
faiss-cpu